Split seq crate into multiple mods. Moved tests into crate.
