+++ crates/seq/src/lib.rs
@@ -1,5 +1,8 @@
 use log::{debug, info};
 use std::fmt::Display;
+pub mod flow;
+pub mod logging;
+pub mod sequencer;
 
 pub trait Node {
     fn execute(&mut self, _delta: f64) -> bool {
@@ -13,30 +16,6 @@ pub trait Node {
     }
     fn exit(&self) {
         // Override
-    }
-}
-
-pub struct SeqLog {
-    text: String,
-}
-
-impl SeqLog {
-    pub fn create(text: &str) -> Box<Self> {
-        Box::new(SeqLog {
-            text: text.to_owned(),
-        })
-    }
-}
-
-impl Display for SeqLog {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.text)
-    }
-}
-
-impl Node for SeqLog {
-    fn enter(&mut self) {
-        info!("SeqLog: {}", self);
     }
 }
 
@@ -141,138 +120,5 @@ impl Node for SeqCheckpoint {
 
     fn advance_to_checkpoint(&mut self, checkpoint: &str) -> bool {
         self.checkpoint_name == checkpoint
-    }
-}
-
-pub struct SeqIf<T: SeqCondition> {
-    name: String,
-    on_true: Option<Box<dyn Node>>,
-    on_false: Option<Box<dyn Node>>,
-    condition: T,
-    selection: bool,
-    default_selection: bool,
-}
-
-impl<T: SeqCondition> SeqIf<T> {
-    pub fn create(
-        name: &str,
-        condition: T,
-        on_true: Option<Box<dyn Node>>,
-        on_false: Option<Box<dyn Node>>,
-        default_selection: bool,
-    ) -> Box<Self> {
-        Box::new(SeqIf {
-            name: name.to_owned(),
-            condition,
-            on_true,
-            on_false,
-            selection: false,
-            default_selection,
-        })
-    }
-}
-
-impl<T: SeqCondition> Node for SeqIf<T> {
-    // When first entering the node, evaluate the conditional
-    fn enter(&mut self) {
-        self.selection = self.condition.evaluate();
-        info!("SeqIf({}), selecting path: {}", self.name, self.selection);
-        match self.selection {
-            true => {
-                if let Some(child) = &mut self.on_true {
-                    child.enter();
-                }
-            }
-            false => {
-                if let Some(child) = &mut self.on_false {
-                    child.enter();
-                }
-            }
-        }
-    }
-    // Execute the selected path until it terminates
-    fn execute(&mut self, delta: f64) -> bool {
-        match self.selection {
-            true => {
-                if let Some(child) = &mut self.on_true {
-                    return child.execute(delta);
-                }
-                true
-            }
-            false => {
-                if let Some(child) = &mut self.on_false {
-                    return child.execute(delta);
-                }
-                true
-            }
-        }
-    }
-    // If advancing past checkpoint, select the default path
-    // TODO: Select based on data instead?
-    fn advance_to_checkpoint(&mut self, checkpoint: &str) -> bool {
-        self.selection = self.default_selection;
-        match self.selection {
-            true => {
-                if let Some(child) = &mut self.on_true {
-                    return child.advance_to_checkpoint(checkpoint);
-                }
-                false
-            }
-            false => {
-                if let Some(child) = &mut self.on_false {
-                    return child.advance_to_checkpoint(checkpoint);
-                }
-                false
-            }
-        }
-    }
-    fn exit(&self) {
-        match self.selection {
-            true => {
-                if let Some(child) = &self.on_true {
-                    child.exit();
-                }
-            }
-            false => {
-                if let Some(child) = &self.on_false {
-                    child.exit();
-                }
-            }
-        }
-    }
-}
-
-pub trait SeqCondition {
-    // TODO(orkaboy): Override. Also needs state/data here!
-    fn evaluate(&self) -> bool {
-        false
-    }
-}
-
-pub struct Sequencer {
-    root: Box<dyn Node>,
-}
-
-impl Sequencer {
-    pub fn create(root: Box<dyn Node>) -> Self {
-        Sequencer { root }
-    }
-}
-
-impl Sequencer {
-    pub fn run(&mut self) {
-        let mut timer = delta::Timer::new();
-        self.root.enter();
-        loop {
-            let deltatime = timer.mark_secs();
-            if self.root.execute(deltatime) {
-                break;
-            }
-        }
-        self.root.exit();
-    }
-
-    pub fn advance_to_checkpoint(&mut self, checkpoint: &str) -> bool {
-        self.root.advance_to_checkpoint(checkpoint)
     }
 }
